Remove donutkata Java 8 version. Refactor donotkata10 package name.
