+++ donut-kata/src/main/java/bnymellon/codekatas/donutkata/Order.java
@@ -18,28 +18,22 @@ package bnymellon.codekatas.donutkata;
 
 import java.time.LocalDate;
 
-import org.eclipse.collections.api.bag.Bag;
 import org.eclipse.collections.api.bag.MutableBag;
 import org.eclipse.collections.api.tuple.primitive.ObjectIntPair;
 import org.eclipse.collections.impl.factory.Bags;
 import org.eclipse.collections.impl.list.fixed.ArrayAdapter;
 import org.eclipse.collections.impl.tuple.primitive.PrimitiveTuples;
 
-public class Order
+public record Order(Customer customer, LocalDate date, MutableBag<DonutType> counts)
 {
-    private Customer customer;
-    private LocalDate date;
-    private MutableBag<DonutType> counts = Bags.mutable.empty();
-
     public Order(Customer customer, LocalDate date, String donutTypeCounts)
     {
-        this.customer = customer;
+        this(customer, date, Bags.mutable.empty());
         ArrayAdapter.adapt(donutTypeCounts.split(","))
                 .asLazy()
                 .collect(pair -> pair.split(":"))
                 .collect(pair -> PrimitiveTuples.pair(DonutType.forAbbreviation(pair[0]), Integer.parseInt(pair[1])))
                 .each(this::add);
-        this.date = date;
     }
 
     private void add(ObjectIntPair<DonutType> pair)
@@ -47,22 +41,12 @@ public class Order
         this.counts.addOccurrences(pair.getOne(), pair.getTwo());
     }
 
-    public Customer getCustomer()
-    {
-        return this.customer;
-    }
-
-    public Bag<DonutType> getCounts()
-    {
-        return this.counts.asUnmodifiable();
-    }
-
-    public LocalDate getDate()
-    {
-        return this.date;
-    }
-
     @Override
+    public int hashCode()
+    {
+        return System.identityHashCode(this);
+    }
+
     public String toString()
     {
         return "Order(" +
