+++ donut-kata/src/test/java/bnymellon/codekatas/donutkata/DonutShopTest.java
@@ -20,18 +20,15 @@ import java.time.Clock;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.ZoneOffset;
-import java.util.DoubleSummaryStatistics;
 
-import org.eclipse.collections.api.list.MutableList;
-import org.eclipse.collections.api.multimap.Multimap;
-import org.eclipse.collections.api.tuple.primitive.ObjectIntPair;
 import org.eclipse.collections.impl.factory.Lists;
 import org.eclipse.collections.impl.test.Verify;
-import org.eclipse.collections.impl.tuple.primitive.PrimitiveTuples;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+
+import static org.eclipse.collections.impl.tuple.primitive.PrimitiveTuples.pair;
 
 public class DonutShopTest
 {
@@ -54,16 +51,16 @@ public class DonutShopTest
         this.donutShop.makeDonuts(DonutType.JELLY, 10);
         this.donutShop.makeDonuts(DonutType.VANILLA_FROSTED, 10);
 
-        Delivery delivery1 =
+        var delivery1 =
                 this.donutShop.deliverOrder("Ted Smith", this.today, "BC:2,BA:1,B:2");
         Assert.assertEquals(6.75d, delivery1.getTotalPrice(), 0.001);
-        Delivery delivery2 =
+        var delivery2 =
                 this.donutShop.deliverOrder("Mary Williams", this.today, "BC:1,G:1");
         Assert.assertEquals(2.70d, delivery2.getTotalPrice(), 0.001);
-        Delivery delivery3 =
+        var delivery3 =
                 this.donutShop.deliverOrder("Sally Prince", this.tomorrow, "BC:6,P:2,B:2,OF:2");
         Assert.assertEquals(12.0d, delivery3.getTotalPrice(), 0.001);
-        Delivery delivery4 =
+        var delivery4 =
                 this.donutShop.deliverOrder("Donnie Dapper", this.yesterday, "BC:6,P:2,B:2,OF:2,G:10");
         Assert.assertEquals(20.9d, delivery4.getTotalPrice(), 0.001);
 
@@ -79,10 +76,9 @@ public class DonutShopTest
     @Test
     public void getTop2Donuts()
     {
-        MutableList<ObjectIntPair<DonutType>> expected =
-                Lists.mutable.with(
-                        PrimitiveTuples.pair(DonutType.BOSTON_CREAM, 15),
-                        PrimitiveTuples.pair(DonutType.GLAZED, 11));
+        var expected = Lists.mutable.empty()
+                .with(pair(DonutType.BOSTON_CREAM, 15))
+                .with(pair(DonutType.GLAZED, 11));
         Assert.assertEquals(expected, this.donutShop.getTopDonuts(2));
     }
 
@@ -106,13 +102,14 @@ public class DonutShopTest
     @Test
     public void getTopCustomer()
     {
-        Assert.assertEquals("Donnie Dapper", this.donutShop.getTopCustomer().getName());
+        Assert.assertEquals("Donnie Dapper", this.donutShop.getTopCustomer().name());
     }
 
     @Test
     public void getCustomersByDonutTypesOrdered()
     {
-        Multimap<DonutType, Customer> multimap = this.donutShop.getCustomersByDonutTypesOrdered();
+        // Do you know the type of multimap?
+        var multimap = this.donutShop.getCustomersByDonutTypesOrdered();
         Assert.assertEquals(6, multimap.keySet().size());
         Verify.assertIterableSize(1, multimap.get(DonutType.BAVARIAN_CREAM));
         Verify.assertAllSatisfy(
@@ -123,26 +120,22 @@ public class DonutShopTest
     @Test
     public void getDonutPriceStatistics()
     {
-        DoubleSummaryStatistics stats1 =
-                this.donutShop.getDonutPriceStatistics(this.today, this.today);
+        var stats1 = this.donutShop.getDonutPriceStatistics(this.today, this.today);
         Assert.assertEquals(9.45d, stats1.getSum(), 0.01);
         Assert.assertEquals(1.35d, stats1.getAverage(), 0.01);
         Assert.assertEquals(7, stats1.getCount(), 0.01);
 
-        DoubleSummaryStatistics stats2 =
-                this.donutShop.getDonutPriceStatistics(this.tomorrow, this.tomorrow);
+        var stats2 = this.donutShop.getDonutPriceStatistics(this.tomorrow, this.tomorrow);
         Assert.assertEquals(12.0d, stats2.getSum(), 0.01);
         Assert.assertEquals(1.0d, stats2.getAverage(), 0.01);
         Assert.assertEquals(12, stats2.getCount(), 0.01);
 
-        DoubleSummaryStatistics stats3 =
-                this.donutShop.getDonutPriceStatistics(this.yesterday, this.yesterday);
+        var stats3 = this.donutShop.getDonutPriceStatistics(this.yesterday, this.yesterday);
         Assert.assertEquals(20.9d, stats3.getSum(), 0.001);
         Assert.assertEquals(0.95d, stats3.getAverage(), 0.01);
         Assert.assertEquals(22, stats3.getCount(), 0.01);
 
-        DoubleSummaryStatistics statsTotal =
-                this.donutShop.getDonutPriceStatistics(this.yesterday, this.tomorrow);
+        var statsTotal = this.donutShop.getDonutPriceStatistics(this.yesterday, this.tomorrow);
         Assert.assertEquals(42.35d, statsTotal.getSum(), 0.01);
         Assert.assertEquals(1.03d, statsTotal.getAverage(), 0.01);
         Assert.assertEquals(41, statsTotal.getCount(), 0.01);
